python_service/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-dotenv==1.0.1
pydantic==2.9.0

langchain==0.3.0
langchain-core
langchain-groq==0.2.0
langchain-community==0.3.0

langgraph==0.2.50
langgraph-checkpoint-postgres==2.0.8

psycopg==3.2.3
psycopg-pool==3.2.4

python-dateutil==2.9.0

sentence-transformers==3.0.1
numpy==1.26.4